@@ -5,19 +5,19 @@
   "dependencies": {
     "@stripe/react-stripe-js": "^1.4.0",
     "@stripe/stripe-js": "^1.14.0",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
+    "@testing-library/jest-dom": "5.14.1",
+    "@testing-library/react": "13.0.0",
+    "@testing-library/user-event": "13.2.1",
     "axios": "^0.21.2",
     "js-cookie": "^2.2.1",
-    "rc-slider": "^9.7.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "rc-slider": "^10.0.1",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-dropzone": "^11.3.2",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
+    "react-scripts": "5.0.1",
     "react-spinners-kit": "^1.9.1",
-    "web-vitals": "^1.0.1"
+    "web-vitals": "2.1.0"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -26,7 +26,10 @@
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
-    "extends": "react-app"
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
   },
   "browserslist": {
     "production": [
